+++ scripts/bash_aliases_argmap.sh
@@ -7,22 +7,16 @@ reset_repo() {
   echo 'Restoring output folder to match remote.'
   git checkout -- "$WORKSPACE/Output/"
   git checkout -- "$WORKSPACE/examples/"
-  rm "$MJS_WP_MAP"
 }
 
 clean_repo() {
   rm "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
   rm "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
-  rm "$MJS_WP_MAP"
 }
 
 # Checks `src/`` for lua files with leftover test/debug code.
 check_repo() {
   grep -Frni "$WORKSPACE/src" -e 'logger:setLevel(logging.DEBUG)' -e 'require("lldebugger").start()'
-}
-
-mappack() {
-  npm --prefix "$MJS_WP_HOME" run pack-js
 }
 
 save_env() {
@@ -32,10 +26,13 @@ save_env() {
 }
 
 # lua argmap2mup Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml > Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup
+# TODO add option for .mup vs .json output
 a2m() {                                    # a2m Input/WhiteSwan_minus_inline_objections_restoring.yml
   NAME=$(basename --suffix=".yml" "$1") && # && ensures error failure stops remaining commands.
-    argmap2mup "$1" >"${2:-$WORKSPACE/Output/$NAME.mup}" &&
-    echo "Generated: ${2:-$WORKSPACE/Output/$NAME.mup}"
+    OUTPUT=${2:-$WORKSPACE/Output/$NAME.mup} &&
+    argmap2mup "$1" >"$OUTPUT" &&
+    # TODO: Should return $2 value so can be used by calling app e.g. a2mo or a2mj.
+    echo "Generated: $OUTPUT"
 }
 
 a2mu() {
@@ -44,16 +41,37 @@ a2mu() {
     echo "Uploaded: $1 to GDrive."
 }
 
+#TODO add way to disable redirect to /dev/null
+pack_mapjs() {
+  # OUTPUT=$1 # Remove default so can test properly: ${1:-$MJS_WP_MAP}
+  # First -- ensures rest is passed onto webpack call
+  # TODO - adding --inspect should enable debug mode - but can't get to work.
+  npm --prefix "$MJS_WP_HOME" run pack-js -- --env.input_map="$1"
+}
+
+# LEGACY, use a2jo instead.
 a2mo() {
-  a2m "$1" "${2:-$MJS_WP_MAP}" &&
-    mappack >/dev/null
-  chrome "$MJS_WP_HOME/index.html" --window-size=500,720 2>/dev/null &
+  NAME=$(basename --suffix=".yml" "$1") &&
+    OUTPUT=${2:-$WORKSPACE/Output/$NAME.json} &&
+    a2m "$1" "$OUTPUT" &&
+    pack_mapjs "$OUTPUT"
+  google-chrome --no-default-browser-check --window-size=500,720 "$MJS_WP_HOME/index.html" 2>/dev/null &
+}
+
+a2jo() {
+  NAME=$(basename --suffix=".yml" "$1")
+  OUTPUT=${2:-$WORKSPACE/Output/$NAME.json}
+  #TODO: Is there a way to pipe a2m output to pack_mapjs?
+  a2m "$1" "$OUTPUT" &&
+    pack_mapjs "$OUTPUT" &&
+    google-chrome --no-default-browser-check --window-size=500,720 "$MJS_WP_HOME/index.html" 2>/dev/null &
 }
 
 m2a() { # m2a Output/Example1_simple.mup
-  NAME=$(basename --suffix=".mup" "$1") &&
-    mup2argmap "$1" >"${2:-$WORKSPACE/Output/$NAME.yml}" &&
-    echo "Generated: ${2:-$WORKSPACE/Output/$NAME.yml}"
+  OUTPUT=${2:-$WORKSPACE/Output/$NAME.yml}
+  NAME=$(basename --suffix=".mup" "$1")
+  mup2argmap "$1" >"$OUTPUT" &&
+    echo "Generated: $OUTPUT"
 }
 
 a2t() {
@@ -77,7 +95,7 @@ md2pdf() { # md2h Input/example.md
 }
 
 ## Mark functions for export to use in other scripts:
-export -f reset_repo clean_repo check_repo mappack save_env a2m m2a a2t a2mu a2mo md2htm md2pdf
+export -f reset_repo clean_repo check_repo pack_mapjs save_env a2m m2a a2t a2mu a2mo a2jo md2htm md2pdf
 
 # argmap Aliases
 
@@ -89,9 +107,11 @@ alias argmm='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup
 alias argmy='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.yml; m2a $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.mup'
 alias argmt='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.tex; a2t $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml'
 alias argmu='a2mu $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml'
-alias argmo='rm $MJS_WP_MAP; a2mo $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml'
+
+#TODO: need to delete previous file, best way? Separate output folder or just delete all .json in Output folder?
+alias argmo='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup; a2mo $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml'
+alias argmj='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.json; a2jo $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml'
 alias argmh='rm $WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.html; rm $WORKSPACE/12ff0311ebc308e94fe0359b761fa405b605f126.png; rm $WORKSPACE/Output/12ff0311ebc308e94fe0359b761fa405b605f126.png; md2htm $WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.md'
 alias argmp='rm $WORKSPACE/Output/example.pdf; md2pdf $WORKSPACE/examples/example.md'
 alias argmph='rm $WORKSPACE/Output/example.pdf; rm $WORKSPACE/Output/header.tex; $WORKSPACE/src/argmap2tikz.lua -i > $WORKSPACE/Output/header.tex; pandoc $WORKSPACE/examples/example.md -o $WORKSPACE/Output/example.pdf --lua-filter pandoc-argmap.lua --pdf-engine lualatex --include-in-header $WORKSPACE/Output/header.tex --data-dir=$CONDA_PREFIX/share/pandoc; echo "Generated: $WORKSPACE/Output/example.pdf"'
-
 alias argt='$WORKSPACE/scripts/tests.sh'